Adds useful types to new nested siteinfotest package
Several types that were previously defined in client_test.go are now
located in a new nested package named "siteinfotest". This makes these
useful types accessible by other modules that may import this package
and and need them for testing.
